refactor config model + add environment options
